RearrangingArray.java
Given a list of integers, rearrange the list such that it consists of alternating minimum-maximum elements using only list operations. The first element of the list should be the minimum of all elements and the second element should be a maximum of all elements present in the list. Similarly, the third element will be the next minimum element and the fourth element is the next maximum element, and so on. Use of extra space is not permitted. Store the answer in the answer[] array. You don't need to print that.

Note : All the elements in the array are unique.

Example 1:

Input
5
4 5 1 2 3
Output
1 5 2 4 3